+++ workspace/solutions/ecmascript/src/s0/s1/s9/s1/s_0191_00.js
@@ -1,0 +1,22 @@
+/*
+ * @lc app=leetcode.cn id=191 lang=javascript
+ *
+ * [191] 位1的个数
+ */
+
+// @lc code=start
+/**
+ * @param {number} n
+ * @return {number}
+ */
+const hammingWeight = function (n) {
+    let count = 0;
+    while (n) {
+        n &= n - 1;
+        count++;
+    }
+    return count;
+};
+// @lc code=end
+
+export default hammingWeight;