+++ workspace/solutions/ecmascript/src/s0/s1/s9/s1/s_0191_test.ts
@@ -1,0 +1,8 @@
+import test from "bun:test";
+
+import { t } from "@/utils/test";
+
+t(
+    test,
+    import.meta.dir,
+);