+++ workspace/solutions/kotlin/lib/src/test/kotlin/s0/s1/s9/s1/s_0191_test.kt
@@ -1,0 +1,18 @@
+package s0.s1.s9.s1
+
+import utils.T
+import kotlin.test.Test
+import s0.s1.s9.s1.s_0191_00.Solution as s_0191_00
+import s0.s1.s9.s1.s_0191_01.Solution as s_0191_01
+
+class s_0191_test {
+    @Test
+    fun test() {
+        val t = T(
+            this::class,
+            s_0191_00::class,
+            s_0191_01::class,
+        )
+        t.run()
+    }
+}