+++ workspace/solutions/rust/src/s0/mod.rs
@@ -1,4 +1,5 @@
 pub mod s0;
+pub mod s1;
 pub mod s3;
 pub mod s4;
 pub mod s7;