+++ workspace/solutions/rust/src/s0/s1/s9/s1/mod.rs
@@ -1,0 +1,6 @@
+pub mod s_0191_00;
+pub mod s_0191_01;
+pub mod s_0191_02;
+pub mod s_0191_test;
+
+pub type TSolution = fn(i32) -> i32;